requirements.txt
streamlit>=1.36.0
openai>=1.30.0
moviepy>=1.0.3
Pillow>=10.3.0
numpy>=1.26.4
